Move get_players_sorted test
