+++ tests/ladder_tests/base_ladder_test.py
@@ -139,9 +139,37 @@ def test_run_game():
            (player1.num_losses == 1 and player2.num_losses == 0))
 
 
+def test_get_players_sorted():
+    """Run get_players with sorted flag to true."""
+    # Set up variables
+    ba1 = BaseAgent()
+    ba2 = BaseAgent()
+    cfe = CoinFlipEngine()
+    lad = BaseLadder(game=cfe)
+    lad.match_func = mock_match_func
+
+    # Add players to the ladder
+    lad.add_player(ba1)
+    lad.add_player(ba2)
+
+    # Run the game
+    lad.run_game()
+
+    # Check that the results are sorted in ascending elo
+    players = lad.get_players(sort=True)
+
+    player1 = players[0]
+    player2 = players[1]
+
+    assert player1.elo > player2.elo
+    assert (player1.num_wins == 1 and player2.num_wins == 0)
+    assert (player1.num_losses == 0 and player2.num_losses == 1)
+
+
 test_add()
 test_duplicate_add()
 test_available_players()
 test_match_basic()
 test_match_error()
 test_run_game()
+test_get_players_sorted()
